fix: update attackUtils imports after file relocation
The rebase onto dev picked up the move of attackUtils.ts from
src/view/sheets/components/ to src/utils/. Two import paths
in ItemActivationManager.ts and DamageRoll.test.ts needed updating
to match the new location.
